refactor(status-transition): share grant form and post controller between accreditation and registration (PAE-1599)
