+++ src/server/routes/registration-status-transition/grant-form.js
@@ -1,28 +1,4 @@
-/**
- * Builds a YYYY-MM-DD date from GDS date-input parts, or null when the parts
- * do not form a real calendar date.
- * @param {string} day
- * @param {string} month
- * @param {string} year
- * @returns {string | null}
- */
-const toIsoDate = (day, month, year) => {
-  if (
-    !/^\d{1,2}$/.test(day) ||
-    !/^\d{1,2}$/.test(month) ||
-    !/^\d{4}$/.test(year)
-  ) {
-    return null
-  }
-
-  const iso = `${year}-${month.padStart(2, '0')}-${day.padStart(2, '0')}`
-  const date = new Date(`${iso}T00:00:00.000Z`)
-  if (Number.isNaN(date.getTime()) || !date.toISOString().startsWith(iso)) {
-    return null
-  }
-
-  return iso
-}
+import { parseAppliesFromDate } from '#server/common/helpers/status-transition/applies-from-date.js'
 
 /**
  * Parses and validates the grant fields (applies from date parts and
@@ -35,17 +11,19 @@ const toIsoDate = (day, month, year) => {
  * }}
  */
 export const parseGrantForm = (payload) => {
-  const day = (payload['appliesFrom-day'] ?? '').trim()
-  const month = (payload['appliesFrom-month'] ?? '').trim()
-  const year = (payload['appliesFrom-year'] ?? '').trim()
+  const {
+    day,
+    month,
+    year,
+    appliesFrom,
+    error: appliesFromError
+  } = parseAppliesFromDate(payload)
   const registrationNumber = (payload.registrationNumber ?? '').trim()
-
-  const appliesFrom = toIsoDate(day, month, year)
 
   /** @type {{ appliesFrom?: string, registrationNumber?: string }} */
   const errors = {}
-  if (!appliesFrom) {
-    errors.appliesFrom = 'Enter a valid applies from date'
+  if (appliesFromError) {
+    errors.appliesFrom = appliesFromError
   }
   if (!registrationNumber) {
     errors.registrationNumber = 'Enter a registration number'
